chore(subscriptions): move frontend code into products/subscriptions
Relocate the insight/dashboard subscriptions frontend (scenes, components, logics) out of frontend/src into the products/subscriptions product, and register its scenes/routes/urls via a product manifest instead of the core scene files.

- Move frontend/src/scenes/subscriptions and frontend/src/lib/components/Subscriptions into products/subscriptions/frontend
- Add products/subscriptions/manifest.tsx (scenes, routes, urls) and drop the corresponding entries from scenes.ts, appScenes.ts, urls.ts
- Add product package.json deps and tsconfig.json so the product typechecks and resolves in isolation
- Update all external importers to the new paths

Generated-By: PostHog Code
Task-Id: e2e0bb2f-d21f-4f81-961e-6d88b57d2980
