+++ products/subscriptions/frontend/components/Subscriptions/InsightSelector.test.tsx
