+++ products/subscriptions/frontend/components/Subscriptions/InsightSelector.tsx
