+++ products/subscriptions/frontend/components/Subscriptions/insightSelectorLogic.test.ts
