+++ products/subscriptions/frontend/scenes/components/SubscriptionDeliveryHistory.tsx
@@ -8,17 +8,18 @@ import {
     LemonTag,
     Tooltip,
 } from '@posthog/lemon-ui'
+
+import { TZLabel } from 'lib/components/TZLabel'
+
 import type {
     PaginatedSubscriptionDeliveryListApi,
     SubscriptionApi,
     SubscriptionDeliveryApi,
-} from '@posthog/products-subscriptions/frontend/generated/api.schemas'
+} from 'products/subscriptions/frontend/generated/api.schemas'
 import {
     SubscriptionDeliveryStatusEnumApi,
     SubscriptionsDeliveriesListStatus as SubscriptionDeliveriesListStatusByValue,
-} from '@posthog/products-subscriptions/frontend/generated/api.schemas'
-
-import { TZLabel } from 'lib/components/TZLabel'
+} from 'products/subscriptions/frontend/generated/api.schemas'
 
 import type { DeliveryFeedback } from '../subscriptionSceneLogic'
 import {
